+++ src/screen/macos/mod.rs
@@ -52,15 +52,13 @@ impl Screen {
 
     /// returns screen dimensions. All monitors included
     pub fn dimension(&self) -> (i32, i32) {
-        let dimensions = (self.screen_width, self.screen_height);
-        dimensions
+        (self.screen_width, self.screen_height)
     }
 
     #[allow(dead_code)]
     /// return region dimension which is set up when template is precalculated
     pub fn region_dimension(&self) -> (u32, u32) {
-        let dimensions = (self.screen_region_width, self.screen_region_height);
-        dimensions
+        (self.screen_region_width, self.screen_region_height)
     }
 
     #[allow(dead_code)]
@@ -138,7 +136,7 @@ impl Screen {
             let gray_value = ((r * 30 + g * 59 + b * 11) / 100) as u8;
             grayscale_data.push(gray_value);
         }
-        let mut image = GrayImage::from_raw(
+        let image = GrayImage::from_raw(
             (self.scaling_factor_x * self.screen_width as f64) as u32,
             (self.scaling_factor_y * self.screen_height as f64) as u32,
             grayscale_data,
@@ -147,7 +145,7 @@ impl Screen {
             "Could not convert image to grayscale".to_string(),
         ))?;
         let image = resize(
-            &mut image,
+            &image,
             self.screen_width as u32,
             self.screen_height as u32,
             Nearest,